requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
pydantic-settings>=2.1.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
twilio>=8.10.0
reportlab>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
httpx>=0.26.0
celery>=5.3.0
redis>=5.0.0

